services/gtfs-rt-archiver/pyproject.toml
[project]
name = "gtfs-rt-archiver"
version = "0.1.0"
description = "Downloads GTFS-RT feeds and uploads them to Google Cloud Storage"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-pubsub>=2.37.0",
    "google-cloud-secret-manager>=2.26.0",
    "google-cloud-storage>=3.10.1",
]

[dependency-groups]
dev = [
    "gtfs-realtime-bindings>=2.0.0",
    "pytest>=9.0.2",
    "pytest-dotenv>=0.5.2",
    "pytest-recording>=0.13.4",
]

[tool.pytest.ini_options]
testpaths = "tests"
env_files = [".env.test"]
